+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 - Add support for libsystemd
 - Add Test for formatSizeValue
+- Allow masking arguments
 - CI Tests: Try more alpine versions
 - Command line option for even more config details
 - Documentation on Return Path Filtering